wip strip submodules
