+++ src/renderer/store/modules/peer/index.js
@@ -1,295 +1,30 @@
-import { isEmpty, random, shuffle } from 'lodash'
-import { PeerDiscovery } from '@arkecosystem/peers'
+
 import ClientService from '@/services/client'
-import config from '@config'
 import i18n from '@/i18n'
 // import PeerModel from '@/models/peer'
-import Vue from 'vue'
+
+import current from './current'
+import seed from './seed'
+import discovery from './discovery'
+import available from './available'
+
+const getBaseUrl = (peer) => `${peer.isHttps ? 'https://' : 'http://'}${peer.ip}:${peer.port}`
 
 const logger = {
   ...console
 }
 
-const errors = {
-  array_zero_length: 'Expected array to return a positive length. Returned 0 instead.',
-  falsy_value: 'Expect value to be truthy, but it is falsy instead.',
-  wrong_type: 'Expect value to be one type, but is some other type.',
-  no_network: 'networkId has the falsy value.'
-}
-
-// Get the base URL from a peer
-const getBaseUrl = (peer) => `${peer.isHttps ? 'https://' : 'http://'}${peer.ip}:${peer.port}`
-
-// Remove a peer from a list based on its IP address.
-const removePeerFromList = (peer, list) => {
-  if (!peer) return list
-  if (!list) return false
-  return list.filter(listed => listed.ip !== peer.ip)
-}
-
-// Return the default network ID based on the profile. The profile object is provided by rootGetters.
-const currentNetworkId = getters => {
-  const profile = getters['session/profile']
-  return profile && profile.networkId
-}
-
-// This is used when optional chaining is not available
-const optionalChaining = (fn, failsafe) => {
-  try {
-    return fn()
-  } catch (err) {
-    return failsafe
-  }
-}
-
-// Return an object containing chain and net from the networkId. Eg: ark.devnet => { chain: ark, net: devnet}
-const chainNetFromNetworkId = networkId => (arr => ({ chain: arr[0], net: arr[1] }))(networkId.split('.'))
-
-// Include peer serialization functions
-
 export default {
   namespaced: true,
 
-  state: {
-    all: {},
-    current: {}
+  modules: {
+    available,
+    current,
+    seed,
+    discovery
   },
 
   getters: {
-    /**
-     * Get all peers for current network.
-     * @param {Boolean} [ignoreCurrent=false] Set if the current peer should be ignored when returning a list of all peers.
-     * @param {string} [networkId=null] Network ID, eg 'ark.devnet'.
-     * @return { Object[] | Null} The peer list.
-     */
-    all: (state, getters, _, rootGetters) => ({ ignoreCurrent = false, networkId = null } = {}) => {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) {
-        logger.error(errors.no_network)
-        return null
-      }
-
-      let peers = optionalChaining(() => state.all[networkId].peers, [])
-
-      if (!peers) {
-        logger.error(errors.falsy_value)
-        return null
-      }
-
-      if (ignoreCurrent) {
-        const currentPeer = getters.current()
-        peers = removePeerFromList(currentPeer, peers)
-      }
-
-      if (peers.length < 1) logger.error(errors.array_zero_length)
-
-      return peers
-    },
-
-    /**
-     * Gets the first peer that matches the IP address.
-     * @param {string} ip The IP address of the peer.
-     * @param {string} [networkId = currentNetworkId] The network that the peer is on.
-     * @return {(Object|undefined|Error)} The peer object. Returns undefined if none is found. Returns an error if no peer is provided.
-     */
-    get: (_, getters) => ({ ip } = {}) => {
-      if (!ip) throw new Error('Unable to find peer: no IP is provided')
-
-      const peers = getters.all()
-
-      if (!peers) {
-        logger.error(errors.falsy_value)
-        return undefined
-      }
-
-      const peer = peers.find(peer => peer.ip === ip)
-
-      if (!peer) logger.error(errors.falsy_value)
-
-      return peer
-    },
-
-    /**
-     * Return the best peer lists. It defaults to RANDOM from TOP 10, but it can be changed.
-     * @param {string} [networkId = currentNetworkId()] The networkId.
-     * @param {bool} [ignoreCurrent = true] Ignore current peer. This filter is applied before the top.
-     * @param {number} [min = 1] Minimum number of peers returned.
-     * @param {number} [max] Maximum number of peers returned.
-     * @param {number} [top = 10] Selection of the N top peers before random is applied
-     */
-    best: (_, getters, __, rootGetters) => ({ networkId, ignoreCurrent = true, min = 1, max, top = 10 } = {}) => {
-      if ((min < 0) || (top < 1)) throw new Error('Impossible parameters on best')
-
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      let peers = getters.all({
-        networkId,
-        ignoreCurrent
-      })
-
-      if (!peers) {
-        logger.error(errors.falsy_value)
-        return []
-      }
-
-      if (!Array.isArray(peers)) logger.error(errors.wrong_type)
-
-      // Prevents minimum and maximum to be bigger than the total of peers.
-      if (min > peers.length) logger.error('Impossible minimum length. Ignoring.')
-
-      max = Math.min(max, peers.length)
-      top = Math.min(top, peers.length)
-
-      const quantity = max ? random(min, max) : peers.length
-
-      // TOP
-      peers = peers.slice(0, top)
-
-      // RANDOM
-      peers = peers.sort(() => 0.5 - Math.random())
-      peers = peers.slice(0, quantity)
-
-      return peers
-    },
-
-    /**
-     * Retrieves an amount of random peers
-     * @param {number} [amount=5] Number of peers to return.
-     * @param {Boolen} [ignoreCurrent = true ] Ignore current peer when returning the random list.
-     * @param {string} networkId Include the network Id.
-     * @return {Object[]} containing peer objects
-     */
-    random: (_, getters) => ({ amount = 5, ignoreCurrent = true, networkId } = {}) => {
-      const peers = getters.all({ ignoreCurrent, networkId })
-
-      return peers.length ? shuffle(peers).slice(0, amount) : []
-    },
-
-    'seed/all': (_, __, ___, rootGetters) => ({ networkId } = {}) => {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return []
-
-      const peers = optionalChaining(() => config.PEERS[networkId], [])
-
-      return peers
-    },
-
-    /**
-     * Retrieves n random seed peers for the current network (excluding current peer)
-     * Note that these peers are currently taken from a config file and will an empty array
-     * custom networks without a corresponding peers file
-     * @param {Number} amount of peers to return
-     * @return {Object[]} containing peer objects, can be length = 0.
-     */
-    'seed/random': (_, getters) => ({ amount = 5, networkId } = {}) => {
-      const peers = getters['seed/all']({ networkId })
-      return shuffle(peers).slice(0, amount)
-    },
-
-    /**
-     * Returns an array of peers that can be used to broadcast a transaction to
-     * Currently this consists of random top 10 peers + 5 random peers.
-     * @return {Object[]} containing peer objects
-     */
-    broadcast: (_, getters) => ({ networkId = null } = {}) => {
-      const peers = []
-
-      // top 10
-      peers.concat(getters.best({
-        max: 10,
-        min: 10,
-        ignoreCurrent: false,
-        networkId
-      }))
-
-      // 5 random
-      peers.concat(getters.random({
-        amount: 5,
-        networkId
-      }))
-
-      // 5 seed random
-      peers.concat(getters['seed/random']({
-        amount: 5,
-        networkId
-      }))
-
-      return peers
-    },
-
-    /**
-     * Get current peer.
-     * @param {string} networkId The ID of the network. This doesn't make much sense, since you cannot be connected to multiple networks.
-     * @return {(Object|boolean)} - false if no current peer
-     */
-    current: (state, getters, __, rootGetters) => ({ networkId = null } = {}) => {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) {
-        logger.error(errors.no_network)
-        return false
-      }
-
-      const currentPeer = state.current[networkId]
-
-      if (isEmpty(currentPeer)) {
-        logger.warn('currentPeer is empty')
-        return false
-      }
-
-      return currentPeer
-    },
-
-    /**
-     * Get last updated date for peer list.
-     * @return {(Date|null|false)}
-     */
-    lastUpdated: (state, _, __, rootGetters) => () => {
-      const networkId = currentNetworkId(rootGetters)
-
-      if (!networkId) return false
-
-      const networkPeers = optionalChaining(() => state.all[networkId].lastUpdated, null)
-
-      return networkPeers
-    },
-
-    /**
-     * Gets a new peer discovery instance. The discovery is based on the following order:
-     *
-     * 1) It checks for the network peers.
-     * 2) It checks for the peer peers.
-     * 3) It checks for the network server peer.
-     * @param {string} [networkId = currentNetworkId()]
-     * @returns {Promise} The instance of the PeerDiscovery.
-     */
-    discovery: (getters, __, ___, rootGetters) => ({ networkId } = {}) => {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return
-
-      /*
-        The application is configured to find peers based on the network names.
-        The default network IDs are based on the chain + net, so it need to be
-        striped to only use the network name (main or devnet). Otherwise, use
-        the network id.
-      */
-      const defaultNetworks = Object.keys(config.PEERS)
-      const net = defaultNetworks.includes(networkId) ? chainNetFromNetworkId(networkId).net : networkId
-
-      // First it checks for the default network peers.
-      if (net) return PeerDiscovery.new({ networkOrHost: net })
-
-      // Then checks for the peers connected.
-      const currentPeer = getters.current()
-      if (currentPeer) return PeerDiscovery.new({ networkOrHost: `${getBaseUrl(currentPeer)}/api/v2/peers` })
-
-      // And then checks for the default network server
-      const networkServer = rootGetters.network[networkId].server
-      return PeerDiscovery.new({ networkOrHost: `${networkServer}/api/v2/peers` })
-    },
 
     /**
      * Gets the current client for the peer.
@@ -297,7 +32,7 @@ export default {
      * @return {Promise} The client
      */
     'peer/client': (getters) => ({ peer }) => {
-      peer = peer || getters.current()
+      peer = peer || getters['current/current']()
 
       if (!peer) return
 
@@ -310,221 +45,7 @@ export default {
 
   },
 
-  mutations: {
-    SET_PEERS (state, { peers, networkId }) {
-      Vue.set(state.all, networkId, {
-        peers: peers,
-        lastUpdated: Date.now()
-      })
-    },
-
-    CLEAR_PEERS (state, { networkId }) {
-      Vue.set(state.all, networkId, {
-        peers: undefined,
-        lastUpdated: undefined
-      })
-    },
-
-    SET_CURRENT_PEER (state, { peer, networkId }) {
-      Vue.set(state.current, networkId, peer)
-    },
-
-    CLEAR_CURRENT_PEER (state, { networkId }) {
-      Vue.set(state.current, networkId, undefined)
-    }
-  },
-
   actions: {
-    /**
-     * Set peers for the network
-     * @param  {Object[]} peers A list of peer objects.
-     * @param  {string} [networkId = currentNetworkId()] A network ID.
-     * @return {void}
-     */
-    'peers/set' ({ rootGetters, commit }, { peers, networkId } = {}) {
-      if (!peers) {
-        logger.error('No peers to set. Send an empty array if this is the desired behaviour.')
-        return
-      }
-
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) {
-        logger.error('Unable to set peers')
-        return
-      }
-
-      commit('SET_PEERS', {
-        peers,
-        networkId
-      })
-    },
-
-    /**
-     * Set current peer for the network
-     * @param {Object} peer The peer. If you want to set an empty peer, send an empty object.
-     * @param {string} [networkId = defaultNetworkId()] A network ID.
-     * @return {void}
-     */
-    async 'current/set' ({ commit, dispatch, rootGetters }, { peer, networkId, update = true } = {}) {
-      if (!peer) {
-        logger.error('No peer was provided to be set as current.')
-        return
-      }
-
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return
-
-      if (update) {
-        // Use the current peer as the base url for requests.
-        this._vm.$client.host = getBaseUrl(peer)
-
-        peer = await dispatch('peer/update', peer)
-
-        // Update the static fees when setting a new peer.
-        // TODO only when necessary (when / before sending) (if no dynamic)
-        await dispatch('transaction/updateStaticFees', null, { root: true })
-      }
-
-      commit('SET_CURRENT_PEER', {
-        peer,
-        networkId: networkId
-      })
-    },
-
-    /**
-     * Clears the peers available to a network.
-     * @param {string} [networkId = currentNetworkId()]
-     * @returns {void}
-     */
-    'peers/clear' ({ commit, rootGetters }, { networkId } = {}) {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return
-
-      commit('CLEAR_PEERS', { networkId: networkId })
-    },
-
-    /**
-     * Clear the current peer.
-     * @param {string} [networkId=currentNetworkId()]
-     * @returns {void}
-     */
-    'current/clear' ({ commit, rootGetters }, { networkId } = {}) {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return
-
-      commit('CLEAR_CURRENT_PEER', { networkId })
-    },
-
-    /**
-     * Refresh peer list.
-     * @param {Object} [networkId=null] The network object.
-     * @return {void}
-     */
-    async 'peers/refresh' ({ getters, dispatch, rootGetters }, { networkId } = {}) {
-      networkId = networkId || currentNetworkId(rootGetters)
-
-      if (!networkId) return
-
-      let peers = []
-
-      try {
-        const peerDiscovery = await getters.discovery({ networkId })
-
-        peerDiscovery
-          .withLatency(300)
-          .sortBy('latency')
-
-        peers = await peerDiscovery
-          .findPeersWithPlugin('core-api', {
-            additional: [
-              'height',
-              'latency'
-            ]
-          })
-
-        if (!peers.length) {
-          peers = await peerDiscovery
-            .findPeersWithPlugin('core-wallet-api', {
-              additional: [
-                'height',
-                'latency'
-              ]
-            })
-        }
-      } catch (error) {
-        logger.error(error)
-        logger.error('Could not refresh peer list:', error)
-        this._vm.$error(i18n.t('PEER.FAILED_REFRESH'))
-      }
-
-      if (!peers.length) {
-        logger.error('No peers retrieved')
-        this._vm.$error(i18n.t('PEER.FAILED_REFRESH'))
-      }
-
-      dispatch('peers/set', { peers })
-    },
-
-    /**
-     * Get best peer for current network.
-     * @param  {Boolean} [refresh=true] Refresh peer list before finding the best peer.
-     * @param  {Object} [network=null] The network object
-     * @return {(Object|null)}
-     */
-    async 'peers/findBest' ({ dispatch, getters, rootGetters }, { refresh = true, network = {} }) {
-      const networkId = network.id || currentNetworkId(rootGetters)
-
-      if (refresh) {
-        try {
-          await dispatch('peers/refresh', { networkId })
-        } catch (error) {
-          logger.error(error)
-          this._vm.$error(`${i18n.t('PEER.FAILED_REFRESH')}: ${error.message}`)
-        }
-      }
-
-      let peer = network ? getters.best({ ignoreCurrent: true, networkId, min: 1, max: 1 })[0] : getters.best({ min: 1, max: 1 })[0]
-
-      if (!peer) return null
-
-      peer = await dispatch('peer/update', peer)
-
-      return peer
-    },
-
-    /**
-     * Update to the best peer for current network.
-     * @param  {Boolean} [refresh=true] - Should refresh the list when trying to find the best peer.
-     * @param  {Boolean} [skipIfCustom=true] Ignore if a custom peer is set.
-     * @return {(Object|null)}
-     */
-    async 'peers/connectToBest' ({ dispatch, getters }, { refresh = true, skipIfCustom = true }) {
-      if (skipIfCustom) {
-        const currentPeer = getters.current()
-        if (!isEmpty(currentPeer) && currentPeer.isCustom) {
-          // TODO only when necessary (when / before sending) (if no dynamic)
-          await dispatch('transaction/updateStaticFees', null, { root: true })
-
-          return null
-        }
-      }
-
-      let peer
-
-      try {
-        const peer = await dispatch('peers/findBest', { refresh })
-        await dispatch('current/set', { peer })
-      } catch (error) {
-        logger.error(error)
-        if (skipIfCustom) await dispatch('system/clear')
-      }
-
-      return peer
-    },
 
     /**
      * Make sure that the peer is valid for the current network.
@@ -538,15 +59,6 @@ export default {
       if (networkConfig.nethash !== rootGetters['session/network'].nethash) {
         throw new Error('Wrong network')
       }
-    },
-
-    /**
-     * Fallback to seed peer, cleaning all the peer data.
-     */
-    async 'system/clear' ({ dispatch }) {
-      dispatch('peers/clear')
-      dispatch('current/clear')
-      await dispatch('peers/connectToBest', { skipIfCustom: false })
     },
 
     /**
@@ -564,7 +76,7 @@ export default {
           .fetchNetworkConfig(getBaseUrl(peer))
           .nethash
       } catch (err) {
-        console.error('Could not get network config:' + err)
+        logger.error('Could not get network config:' + err)
       }
 
       if (!peerNethash) {
@@ -583,6 +95,7 @@ export default {
      */
     async 'peer/update' ({ dispatch }, peer) {
       let response
+
       try {
         response = await dispatch('peer/fetchStatus', peer)
       } catch (err) {
@@ -595,12 +108,12 @@ export default {
         throw i18n.t('PEER.STATUS_CHECK_FAILED')
       }
 
-      const { status, heigth, latency } = response
+      const { status, height, latency } = response
 
       peer = {
         ...peer,
         status,
-        heigth,
+        height,
         latency
       }
 
@@ -630,28 +143,6 @@ export default {
         status: 'OK',
         height: status.height,
         latency: (latencyEnd - latencyStart).toFixed(0)
-      }
-    },
-
-    /**
-     * Updates the whole peer system, checking the current peer is still valid and updating status.
-     * @param  {Object} [port]
-     * @return {(Object|void)}
-     */
-    async 'system/update' ({ dispatch, getters }) {
-      let peer = getters.current()
-
-      if (!peer) {
-        await dispatch('system/clear')
-        return
-      }
-
-      try {
-        peer = await dispatch('peer/update', peer)
-        await dispatch('current/set', { peer })
-      } catch (error) {
-        await dispatch('system/clear')
-        await dispatch('refresh')
       }
     },
